Navigation to new page works Now have to build single offer view && fetch data about offer to new page

+++ frontend_android/app/src/main/java/pl/pollub/harnasik/app/data/remote/Offer/OfferService.kt
@@ -12,7 +12,7 @@ import pl.pollub.harnasik.app.data.remote.Offer.dto.OfferResponse
 
 interface OfferService {
 
-    suspend fun getOfferById(id: Long): OfferResponse?
+    suspend fun getOfferById(id: Int): OfferResponse?
 
     suspend fun getAllOffers(): List<OfferResponse>
