+++ frontend_android/app/src/main/java/pl/pollub/harnasik/app/data/remote/Offer/OfferServiceImpl.kt
@@ -12,7 +12,7 @@ class OfferServiceImpl(
     private val client: HttpClient
 ) : OfferService {
 
-    override suspend fun getOfferById(id: Long): OfferResponse? {
+    override suspend fun getOfferById(id: Int): OfferResponse? {
         return try {
             client.get{
                 url(HttpRoutes.OFFER)
